requirements.txt
django >= 1.10, < 1.11
django-admin-sortable2
django-compressor
django-sass-processor
libsass
gunicorn
